feat(drawer): Implement drawer state
Closes #579
