--- src/lib/drawer/constants.ts
@@ -1,3 +1,0 @@
-export const FOCUSABLE_ELEMENTS =
-  'a[href], area[href], input:not([disabled]), select:not([disabled]), textarea:not([disabled]), ' +
-  'button:not([disabled]), iframe, object, embed, [tabindex], [contenteditable]';
