+++ src/lib/drawer/public-api.ts
@@ -1,5 +1,2 @@
 export * from './drawer-module';
 export * from './drawer';
-export * from './permanent/drawer-permanent';
-export * from './temporary/drawer-temporary';
-export * from './persistent/drawer-persistent';
