--- src/theme/extend/drawer/mdc-drawer-permanent.scss
@@ -1,7 +1,0 @@
-.ng-mdc-drawer--permanent--fixed {
-  @include mdc-elevation(1);
-
-  position: fixed;
-  height: 100vh;
-  overflow-y: auto;
-}
